requirements.txt
# Flask ~= 3.1.2       
# python-dotenv ~= 1.2.1   
# google-auth-oauthlib ~= 1.2.1  
# google-auth ~= 2.42.1     
# google-api-python-client ~= 2.185.0    
# requests ~= 2.31.0

Flask==3.0.3
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-auth==2.42.1
google-api-python-client==2.185.0
requests==2.31.0
